refactor `ElGamalCiphertext` and `ElGamalPubkey` to separate submodule
