+++ zk-token-sdk/src/zk_token_elgamal/pod/mod.rs
@@ -1,10 +1,15 @@
-pub use bytemuck::{Pod, Zeroable};
+mod elgamal;
+
 use {
     crate::zk_token_proof_instruction::ProofType,
     base64::{prelude::BASE64_STANDARD, Engine},
     num_traits::{FromPrimitive, ToPrimitive},
     solana_program::instruction::InstructionError,
     std::fmt,
+};
+pub use {
+    bytemuck::{Pod, Zeroable},
+    elgamal::{ElGamalCiphertext, ElGamalPubkey},
 };
 
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Pod, Zeroable)]
@@ -54,44 +59,6 @@ impl TryFrom<PodProofType> for ProofType {
 #[derive(Clone, Copy, Pod, Zeroable, PartialEq, Eq)]
 #[repr(transparent)]
 pub struct CompressedRistretto(pub [u8; 32]);
-
-#[derive(Clone, Copy, Pod, Zeroable, PartialEq, Eq)]
-#[repr(transparent)]
-pub struct ElGamalCiphertext(pub [u8; 64]);
-
-impl fmt::Debug for ElGamalCiphertext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self.0)
-    }
-}
-
-impl fmt::Display for ElGamalCiphertext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl Default for ElGamalCiphertext {
-    fn default() -> Self {
-        Self::zeroed()
-    }
-}
-
-#[derive(Clone, Copy, Default, Pod, Zeroable, PartialEq, Eq)]
-#[repr(transparent)]
-pub struct ElGamalPubkey(pub [u8; 32]);
-
-impl fmt::Debug for ElGamalPubkey {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self.0)
-    }
-}
-
-impl fmt::Display for ElGamalPubkey {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
 
 #[derive(Clone, Copy, Default, Pod, Zeroable, PartialEq, Eq)]
 #[repr(transparent)]
